+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 ------------------
 
 * S1 Pattern likelihood in LowEnergyCuts (#21)
+* Max PMT in S1 (v0) LowEnergyCuts (#15)
 
 0.4.0 (2017-02-24)
 ------------------