+++ HISTORY.rst
@@ -2,6 +2,8 @@
 History
 =======
 
+S2AreaFractionTopCut now can have v3 (v2 still default) with tighter AFT selection (#14)
+
 0.4.0 (2017-02-24)
 ------------------